+++ lib/imports.dart
@@ -1,13 +1,16 @@
+export 'dart:convert';
+
 export 'package:flutter/material.dart';
 export 'package:flutter/services.dart';
 export 'package:fluttertoast/fluttertoast.dart';
-export 'dart:convert';
-export 'dart:io';
 export 'package:shared_preferences/shared_preferences.dart';
 export 'package:provider/provider.dart';
+export 'package:connectivity_plus/connectivity_plus.dart';
+
 export 'providers/host_list_provider.dart';
 export 'widgets/add_host.dart';
-export 'widgets/host_name_text_field.dart';
 export 'widgets/ip_address_text_field.dart';
 export 'widgets/mac_address_text_field.dart';
-
+export 'widgets/host_name_text_field.dart';
+export '../utils/utils.dart';
+export '../about.dart';
