+++ lib/widgets/ip_address_text_field.dart
@@ -7,6 +7,7 @@ class IPAddressTextField extends StatefulWidget {
   const IPAddressTextField({Key? key, this.onChanged, required this.ipAddress}) : super(key: key);
 
   @override
+  
   State<IPAddressTextField> createState() => _IPAddressTextFieldState();
 }
 
@@ -32,11 +33,11 @@ class _IPAddressTextFieldState extends State<IPAddressTextField> {
         widget.onChanged?.call(value);
       },
       controller: _controller,
-      keyboardType: TextInputType.numberWithOptions(signed: false, decimal: true),
+      keyboardType: const TextInputType.numberWithOptions(signed: false, decimal: true),
       inputFormatters: [
         MyInputFormatters.ipAddressInputFilter(),
         LengthLimitingTextInputFormatter(15),
-        IpAddressInputFormatter()
+        _IPAddressInputFormatter(), // Using the custom IP address input formatter
       ],
       decoration: const InputDecoration(
         labelText: 'IP Address',
@@ -45,59 +46,18 @@ class _IPAddressTextFieldState extends State<IPAddressTextField> {
   }
 }
 
-class IpAddressInputFormatter extends TextInputFormatter {
+
+class _IPAddressInputFormatter extends TextInputFormatter {
   @override
   TextEditingValue formatEditUpdate(TextEditingValue oldValue, TextEditingValue newValue) {
-    var text = newValue.text;
-
-    if (newValue.selection.baseOffset == 0) {
-      return newValue;
-    }
-
-    int dotCounter = 0;
-    var buffer = StringBuffer();
-    String ipField = "";
-
-    for (int i = 0; i < text.length; i++) {
-      if (dotCounter < 4) {
-        if (text[i] != ".") {
-          ipField += text[i];
-          if (ipField.length < 3) {
-            buffer.write(text[i]);
-          } else if (ipField.length == 3) {
-            if (int.parse(ipField) <= 255) {
-              buffer.write(text[i]);
-            } else {
-              if (dotCounter < 3) {
-                buffer.write(".");
-                dotCounter++;
-                buffer.write(text[i]);
-                ipField = text[i];
-              }
-            }
-          } else if (ipField.length == 4) {
-            if (dotCounter < 3) {
-              buffer.write(".");
-              dotCounter++;
-              buffer.write(text[i]);
-              ipField = text[i];
-            }
-          }
-        } else {
-          if (dotCounter < 3) {
-            buffer.write(".");
-            dotCounter++;
-            ipField = "";
-          }
-        }
-      }
-    }
-
-    var string = buffer.toString();
-    return newValue.copyWith(
-        text: string, selection: TextSelection.collapsed(offset: string.length));
+    final formattedText = formatIPAddress(newValue);
+    return TextEditingValue(
+      text: formattedText,
+      selection: TextSelection.collapsed(offset: formattedText.length),
+    );
   }
 }
+
 
 class MyInputFormatters {
   static TextInputFormatter ipAddressInputFilter() {
