+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   fluttertoast: ^8.2.2
   package_info_plus: ^4.0.2
   url_launcher: ^6.1.11
+  connectivity_plus: ^4.0.2
 
 dev_dependencies:
   flutter_test:
